+++ packages/react-magma-dom/src/components/Modal/Modal.tsx
@@ -185,10 +185,11 @@ const H1 = styled(Heading)<{ theme?: ThemeInterface; isInverse?: boolean }>`
   font-weight: 600;
 `;
 
-const CloseBtn = styled.span`
+const CloseBtn = styled.span<{ theme?: ThemeInterface }>`
   position: absolute;
   top: 0;
   right: 0;
+  margin: ${props => props.theme.spaceScale.spacing02};
 `;
 const ModalBody = styled.div<{ theme?: ThemeInterface }>`
   padding: ${props => props.theme.spaceScale.spacing05};
@@ -379,7 +380,7 @@ export const Modal = React.forwardRef<HTMLDivElement, ModalProps>(
                   {children}
                 </ModalBody>
                 {!isCloseButtonHidden && (
-                  <CloseBtn>
+                  <CloseBtn theme={theme}>
                     <IconButton
                       aria-label={
                         closeAriaLabel