+++ packages/react-magma-dom/src/components/Tooltip/index.tsx
@@ -137,7 +137,7 @@ export const StyledTooltip = styled.div<{
   }
 `;
 
-// Using any for the ref because it is put ont he passed in children which does not have a specific type
+// Using any for the ref because it is put on the passed in children which does not have a specific type
 export const Tooltip = React.forwardRef<any, TooltipProps>((props, ref) => {
   const [isVisible, setIsVisible] = React.useState<boolean>(props.open);
   const [referenceElement, setReferenceElement] =
@@ -215,6 +215,7 @@ export const Tooltip = React.forwardRef<any, TooltipProps>((props, ref) => {
   });
 
   const combinedTooltipStyles = {
+    zIndex: theme.tooltip.zIndex,
     ...styles.popper,
     ...tooltipStyle,
   };