+++ packages/react-magma-dom/src/index.ts
@@ -242,3 +242,4 @@ export * from './components/Grid';
 export * from './components/Tag';
 export * from './components/List';
 export * from './components/BlockQuote';
+export * from './components/ButtonGroup'